temperature_calibration: reduce code duplication by adding a TemperatureCalibrationCommon class
