+++ src/modules/events/temperature_calibration/accel.h
@@ -36,16 +36,11 @@
 #include "common.h"
 #include "polyfit.hpp"
 
-class TemperatureCalibrationAccel : public TemperatureCalibrationBase
+class TemperatureCalibrationAccel : public TemperatureCalibrationCommon<3, 3>
 {
 public:
 	TemperatureCalibrationAccel(float min_temperature_rise);
 	virtual ~TemperatureCalibrationAccel();
-
-	/**
-	 * @see TemperatureCalibrationBase::update()
-	 */
-	int update();
 
 	/**
 	 * @see TemperatureCalibrationBase::finish()
@@ -59,29 +54,7 @@ public:
 
 private:
 
-	struct PerSensorData {
-		float sensor_sample_filt[4];
-		polyfitter<4> P[3];
-		unsigned hot_soak_sat = 0;
-		uint32_t device_id = 0;
-		bool cold_soaked = false;
-		bool hot_soaked = false;
-		bool tempcal_complete = false;
-		float low_temp = 0.f;
-		float high_temp = 0.f;
-		float ref_temp = 0.f;
-	};
-
-	PerSensorData _data[SENSOR_COUNT_MAX];
-
-	/**
-	 * update a single sensor instance
-	 * @return 0 when done, 1 not finished yet
-	 */
-	inline int update_sensor_instance(PerSensorData &data, int sensor_sub);
+	virtual inline int update_sensor_instance(PerSensorData &data, int sensor_sub);
 
 	inline int finish_sensor_instance(PerSensorData &data, int sensor_index);
-
-	int _num_sensor_instances;
-	int _sensor_subs[SENSOR_COUNT_MAX];
 };
